+++ packages/nuxt/src/runtime/server/middleware/federation.ts
@@ -1,0 +1,4 @@
+import { createContextData, federation } from "#fedify/nuxt-config.mjs";
+import { integrateFederation } from "@fedify/h3";
+
+export default integrateFederation(federation, createContextData);